Treat ILP coefficient conversion as an ordinary reduction rule
