+++ src/rules/mod.rs
@@ -43,8 +43,8 @@ pub(crate) mod hamiltoniancircuit_travelingsalesman;
 pub(crate) mod hamiltonianpath_degreeconstrainedspanningtree;
 pub(crate) mod hamiltonianpath_isomorphicspanningtree;
 pub(crate) mod hamiltonianpathbetweentwovertices_longestpath;
-pub(crate) mod ilp_casts;
 pub(crate) mod ilp_i64_ilp_bool;
+pub(crate) mod ilp_i64_ilp_f64;
 pub(crate) mod integerknapsack_ilp;
 pub(crate) mod kclique_balancedcompletebipartitesubgraph;
 pub(crate) mod kclique_conjunctivebooleanquery;
